// types.ts
export interface Product {
  _id: string; 
  name: string;
  slug: {
    current: string;
  };
  srcUrl: {
    asset: {
      _ref: string;
      url: string;
    };
  };
  gallery: {
    asset: {
      _ref: string;
      url: string;
    };
  }[];
  price: number;
  discount: {
    amount: number;
    percentage: number;
  };
  rating: number;
  inStock: boolean;
  quantity?: number;
  stock: number
}

  export interface Review {
    id: number;
    user: string;
    content: string;
    rating: number;
    date: string;
  }

  export type Address = {
    name: string;
    phone: string;
    addressLine1: string;
    addressLine2?: string;
    cityLocality: string;
    stateProvince: string;
    postalCode: string;
    countryCode: string;
    addressResidentialIndicator: "yes" | "no";
  };
  export type unit = "ounce" | "gram" | "kilogram" | "pound";
  export type dimensionUnit = "inch" | "centimeter";
  
  export type Package = {
    weight: {
      value: number;
      unit: unit;
    };
    dimensions: {
      height: number;
      width: number;
      length: number;
      unit: dimensionUnit;
    };
  };
  
  export type Rate = {
    rateId: string;
    rateType: string;
    carrierId: string;
    shippingAmount: {
      currency: string;
      amount: number;
    };
    serviceType: string;
    serviceCode: string;
    trackable: boolean;
    carrierFriendlyName: string;
    validationStatus: string;
    warningMessages?: string[];
  };
  
  export interface trackingObjType {
    trackingNumber: string;
    labelId: string;
    carrierCode: string;
  }
  
  export interface TrackingData {
    trackingNumber?: string;
    statusDescription?: string;
    carrierStatusDescription?: string;
    estimatedDeliveryDate?: string;
    actualDeliveryDate?: string;
  }
  
  export type Discount = {
    amount: number;
    percentage: number;
  };